agent/requirements.txt
langgraph
langchain
langchain-core
langchain-gigachat
python-dotenv
